--- src/main/java/com/example/speedrunnerswap/utils/ActionBarUtil.java
@@ -1,20 +1,0 @@
-package com.example.speedrunnerswap.utils;
-
-import net.md_5.bungee.api.ChatMessageType;
-import net.md_5.bungee.api.chat.TextComponent;
-import org.bukkit.entity.Player;
-
-/**
- * Utility class for sending action bar messages
- */
-public class ActionBarUtil {
-
-    /**
-     * Send an action bar message to a player
-     * @param player The player to send the message to
-     * @param message The message to send
-     */
-    public static void sendActionBar(Player player, String message) {
-        player.spigot().sendMessage(ChatMessageType.ACTION_BAR, TextComponent.fromLegacyText(message));
-    }
-}
