+++ src/main/java/com/yourname/speedrunnerswap/config/Effect.java
@@ -1,0 +1,4 @@
+package com.yourname.speedrunnerswap.config;
+
+public record Effect(boolean enabled, String sound, float volume, float pitch, String title, String subtitle) {
+}
