+++ src/main/java/com/yourname/speedrunnerswap/events/GamePauseEvent.java
@@ -1,0 +1,20 @@
+package com.yourname.speedrunnerswap.events;
+
+import org.bukkit.event.Event;
+import org.bukkit.event.HandlerList;
+import org.jetbrains.annotations.NotNull;
+
+public class GamePauseEvent extends Event {
+
+    private static final HandlerList handlers = new HandlerList();
+
+    @NotNull
+    @Override
+    public HandlerList getHandlers() {
+        return handlers;
+    }
+
+    public static HandlerList getHandlerList() {
+        return handlers;
+    }
+}
