+++ src/main/java/com/yourname/speedrunnerswap/voice/VoiceChatIntegration.java
@@ -1,0 +1,123 @@
+package com.yourname.speedrunnerswap.voice;
+
+import com.yourname.speedrunnerswap.SpeedrunnerSwap;
+import org.bukkit.entity.Player;
+
+/**
+ * Stub for Simple Voice Chat integration.
+ * This class provides a placeholder for integration with the Simple Voice Chat plugin.
+ * When the actual plugin is present, this would be expanded to use its API.
+ */
+public class VoiceChatIntegration {
+    
+    private final SpeedrunnerSwap plugin;
+    private boolean enabled;
+    private boolean pluginDetected;
+    
+    public VoiceChatIntegration(SpeedrunnerSwap plugin) {
+        this.plugin = plugin;
+        this.enabled = plugin.getConfigManager().isVoiceChatIntegrationEnabled();
+        this.pluginDetected = checkForVoiceChatPlugin();
+    }
+    
+    /**
+     * Checks if the Simple Voice Chat plugin is installed and available.
+     * 
+     * @return true if the plugin is detected, false otherwise
+     */
+    private boolean checkForVoiceChatPlugin() {
+        return plugin.getServer().getPluginManager().getPlugin("SimpleVoiceChat") != null;
+    }
+    
+    /**
+     * Mutes a player in voice chat.
+     * This would use the Simple Voice Chat API to mute the player if the plugin is present.
+     * 
+     * @param player The player to mute
+     */
+    public void mutePlayer(Player player) {
+        if (!enabled || !pluginDetected) return;
+        
+        // This is a stub implementation
+        // In a real implementation, this would use the Simple Voice Chat API
+        plugin.getLogger().info("VoiceChat: Would mute player " + player.getName());
+        
+        // Example of how this might be implemented with the actual API:
+        // VoicechatServerApi api = VoicechatServerApi.instance();
+        // api.mutePlayer(player.getUniqueId());
+    }
+    
+    /**
+     * Unmutes a player in voice chat.
+     * This would use the Simple Voice Chat API to unmute the player if the plugin is present.
+     * 
+     * @param player The player to unmute
+     */
+    public void unmutePlayer(Player player) {
+        if (!enabled || !pluginDetected) return;
+        
+        // This is a stub implementation
+        // In a real implementation, this would use the Simple Voice Chat API
+        plugin.getLogger().info("VoiceChat: Would unmute player " + player.getName());
+        
+        // Example of how this might be implemented with the actual API:
+        // VoicechatServerApi api = VoicechatServerApi.instance();
+        // api.unmutePlayer(player.getUniqueId());
+    }
+    
+    /**
+     * Updates the mute status of all runners based on the active runner.
+     * Only the active runner should be unmuted, all other runners should be muted.
+     */
+    public void updateRunnerMuteStatus() {
+        if (!enabled || !pluginDetected) return;
+        
+        Player activeRunner = plugin.getGameManager().getActiveRunner();
+        
+        for (Player runner : plugin.getGameManager().getRunners()) {
+            if (runner.equals(activeRunner)) {
+                unmutePlayer(runner);
+            } else {
+                mutePlayer(runner);
+            }
+        }
+    }
+    
+    /**
+     * Resets the mute status of all players.
+     * This should be called when the game ends to ensure all players are unmuted.
+     */
+    public void resetAllPlayerMuteStatus() {
+        if (!enabled || !pluginDetected) return;
+        
+        for (Player runner : plugin.getGameManager().getRunners()) {
+            unmutePlayer(runner);
+        }
+    }
+    
+    /**
+     * Sets whether voice chat integration is enabled.
+     * 
+     * @param enabled true to enable, false to disable
+     */
+    public void setEnabled(boolean enabled) {
+        this.enabled = enabled;
+        
+        if (!enabled) {
+            // If disabling, make sure to unmute all players
+            resetAllPlayerMuteStatus();
+        } else if (plugin.getGameManager().isGameRunning()) {
+            // If enabling during a running game, update mute status
+            updateRunnerMuteStatus();
+        }
+    }
+    
+    /**
+     * Checks if voice chat integration is enabled.
+     * 
+     * @return true if enabled, false otherwise
+     */
+    public boolean isEnabled() {
+        return enabled && pluginDetected;
+    }
+}
